models/gnn/requirements.txt
tensorflow-gpu==1.14.0
graph_nets==1.0.4
dm-sonnet==1.34
tensorflow-probability==0.7.0
